+++ src/app/(private)/onboarding/character/page.tsx
@@ -1,14 +1,14 @@
 "use client";
 
 import { FormProvider } from "react-hook-form";
+import { usePatchUserCharacter } from "@/services/user";
 import CharacterForm from "./CharacterForm";
 import CharacterPreview from "./CharacterPreview";
 import ColorForm from "./ColorForm";
 import { useCharacterSearchParams } from "./useCharacterSearchParams";
-import { usePatchCharacter } from "./usePatchCharacter";
 
 export default function Page() {
-  const { formMethods, handleSubmit, isPending } = usePatchCharacter();
+  const { formMethods, handleSubmit, isPending } = usePatchUserCharacter();
   const { target, setUrlNameParam, setUrlColorParam } =
     useCharacterSearchParams();
