+++ src/services/family/index.ts
@@ -1,0 +1,1 @@
+export * from "./usePostFamily";
